+++ scripts/qemu-vm.sh
@@ -1083,8 +1083,10 @@ config_vm_tmpdir() {
 	_eval touch ${vm_cmd_sh}
 	_eval chmod +x ${vm_cmd_sh}
 
-	_eval python3 -c "\"open('${vm_port_hostfwd_ssh22}','w').write('${TCP_PORT_HOSTFWM_SSH22}')\""
-	_eval python3 -c "\"open('${vm_port_monitor_telnet}','w').write('${TCP_PORT_MONITOR_TELNET}')\""
+	if [[ -z ${dry_run} ]]; then
+		fprintf ${vm_port_hostfwd_ssh22} ${TCP_PORT_HOSTFWM_SSH22}
+		fprintf ${vm_port_monitor_telnet} ${TCP_PORT_MONITOR_TELNET}
+	fi
 
 	cleanup_files+=( ${vm_cmd_sh} )
 	cleanup_files+=( ${vm_port_hostfwd_ssh22} )